app/requirements.txt
fastapi
uvicorn
sqlalchemy>=1.4.0
alembic>=1.7.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
pydantic
httpx
